+++ library/src/main/java/com/novoda/downloadmanager/lib/DownloadService.java
@@ -40,6 +40,7 @@ import com.novoda.downloadmanager.notifications.CancelledNotificationCustomiser;
 import com.novoda.downloadmanager.notifications.CompleteNotificationCustomiser;
 import com.novoda.downloadmanager.notifications.DownloadingNotificationCustomiser;
 import com.novoda.downloadmanager.notifications.FailedNotificationCustomiser;
+import com.novoda.downloadmanager.notifications.NotificationCustomiser;
 import com.novoda.downloadmanager.notifications.QueuedNotificationCustomiser;
 import com.novoda.downloadmanager.lib.logger.LLog;
